+++ modules/agent/src/main/java/io/jpom/util/FileUtils.java
@@ -123,4 +123,8 @@ public class FileUtils {
         return strings[1];
     }
 
+
+    public static String getJarSeparator() {
+        return SystemUtil.getOsInfo().isWindows() ? ";" : ":";
+    }
 }